improve test code
