+++ test/login/data/api_login/helpers/api_login_service_sut.dart
@@ -1,6 +1,6 @@
 import 'package:thepos/features/login/data/login_service/api_login/api_login_service.dart';
 
-import '../../../helpers/mock_client_stub.dart';
+import '../../../../helpers/mock_client_stub.dart';
 
 class ApiLoginServiceSUT {
   ApiLoginServiceSUT(this.client, this.apiLoginService);
