+++ test/login/data/login_use_case/helpers/login_use_case_sut.dart
@@ -1,0 +1,12 @@
+import 'package:thepos/features/login/data/login_service/login_service.dart';
+import 'package:thepos/features/login/data/login_use_case/login_use_case.dart';
+import 'package:thepos/features/login/data/login_use_case/login_use_case_output.dart';
+
+class LoginUseCaseSUT {
+  LoginUseCaseSUT(
+      {required this.service, required this.output, required this.useCase});
+
+  final LoginService service;
+  final LoginServiceOutput output;
+  final LoginUseCase useCase;
+}
